refactor(role): move from `contents` to `constant`
